First attempt at menu

+++ app/helpers/application_helper.rb
@@ -10,12 +10,11 @@ module ApplicationHelper
   end
 
   def login_label
-    a = if current_user
+    if current_user
       "Hello #{current_user.login} #{link_to 'logout', session_url, :method => :delete}" 
     else
       link_to "Login", new_session_url
     end
-    "<div class='login'>#{a}</div>" 
   end
 
   def edit_link_to(*args)
